# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Business Directory Service API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://business-directory-service.herokuapp.com/api/documentation",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/api/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Sign in",
                "description": "Login by email, password",
                "operationId": "authLogin",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@mail.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "User successfully retrieved"
                                        },
                                        "data": {
                                            "type": "object",
                                            "example": {
                                                "id": 1,
                                                "email": "admin@mail.com",
                                                "created_at": "2021-01-02T15:28:39.000000Z",
                                                "updated_at": "2021-01-02T15:28:39.000000Z",
                                                "token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJleHAiOm51bGwsImlkIjoxLCJlbWFpbCI6ImFkbWluQG1haWwuY29tIn0.y_tl1_zBpO-M3p1Uev-6Uji9AmgQ-zkF1dHi9QZHIgc"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "responses": {
            "422": {
                "description": "Wrong credentials response",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "message": {
                                    "type": "string",
                                    "example": "Sorry, wrong email address or password. Please try again"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}